Upadted the functionality of the game and sorted the game classes by refactoring
