+++ src/game/scenes/levels/Level1.ts
@@ -2,8 +2,11 @@ import Game from '../../../engine/Game.js';
 import Scene from '../../../engine/Scene.js';
 import Player from '../../gameItems/Player.js';
 import FovOverlay from '../../gameItems/FovOverlay.js';
+// import GameItem from '../../../engine/GameItem.js';
+import Level from '../../../engine/Level.js';
+import Input from '../../../engine/Input.js';
 
-export default class Level extends Scene {
+export default class Level1 extends Level {
   // Player Character
   private player: Player;
 
@@ -17,6 +20,40 @@ export default class Level extends Scene {
    */
   public constructor(game: Game) {
     super(game);
+
+    this.input = new Input();
+    /**
+     * Test to see if we can map the gameItems of each level
+     * and call their respective properties and methods
+     */
+    // this.gameItems = new Map<string, GameItem>([
+    //   ['player', new Player(
+    //     // The path to the Source Image of the Player Character appearance
+    //     './assets/img/testplayer-old.png',
+    //     // The coordinates of the Player Character on the game canvas
+    //     game.canvas.width / 2, game.canvas.height / 2,
+    //     // The rotation of the Player Character measured in degrees
+    //     0,
+    //     // The dimensions of the Player Character appearance
+    //     32, 32,
+    //     // The dimensions of the Player Character collider
+    //     32, 32,
+    //     // The current state of the Player Character animation cycle
+    //     0,
+    //   )],
+    //   ['fov', this.fov = new FovOverlay(
+    //     // The path to the Source Image of the Fov appearance
+    //     './assets/img/fov.png',
+    //     // The coordinates of the FovOverlay on the game canvas
+    //     game.canvas.width / 2, game.canvas.height / 2,
+    //     // The rotation of the FovOverlay measured in degrees
+    //     0,
+    //     // The dimensions of the FovOverlay appearance
+    //     6000, 6000,
+    //   )],
+    // ]);
+    // console.log(this.gameItems);
+    // console.log(this.gameItems.get('player'));
 
     // Spawning the Player Character
     this.player = new Player(
@@ -63,9 +100,9 @@ export default class Level extends Scene {
    */
   public update(): Scene {
     // Providing Player Control over the Player Character
-    this.player.control();
+    this.player.control(this.input);
     // Providing Player Control over the FovOverlay
-    this.fov.control();
+    this.fov.control(this.input);
 
     // Preserving the position of the FovOverlay relative to the Player Character
     this.fov.setXPos(this.player.getXPos());
